+++ src/utils.rs
@@ -10,6 +10,8 @@ use tracing::{error, info};
 pub mod batch_type;
 pub mod index_packing;
 pub mod min_map;
+pub mod secret;
+pub mod serde_utils;
 pub mod tree_updates;
 
 pub fn spawn_monitored_with_backoff<S, F>(
